Don't send PATH broadcasts if PATH extensions are disabled
Waiting for the PATH broadcasts to complete can add significantly to
startup time, particularly on Android 4.4+ devices when the device is
busy or just booting up.  If PATH extensions are disabled, we ignore the
results of the broadcasts completely, so we may as well just skip them
in the first place.