Split RUN_SCRIPT handling out into its own activity
We currently use intent filters to direct jackpal.androidterm.RUN_SCRIPT
actions to the RunScript activity, which is gated in the manifest by the
permission jackpal.androidterm.permission.RUN_SCRIPT.  However,
RunScript is an alias for the RemoteInterface activity, which is exposed
without the permission requirement; because applications are permitted to
explicitly specify a component when sending an intent, this allows an
application to trigger a RUN_SCRIPT action without having the
corresponding permission.

Fix this by removing RUN_SCRIPT handling from RemoteInterface and
creating a new RunScript activity (a subclass of RemoteInterface) to
handle RUN_SCRIPT actions.  Any scripts supplied by callers to
RemoteInterface will now be ignored, though the call will open a new
terminal window.

Thanks to Daoyuan Wu for reporting and analyzing the issue.  Fixes #374.