Encrypt shortcut intents
As it turns out, the terminal shortcut feature was relying on the
explicit component intent security hole closed in 5112961... ("Split
RUN_SCRIPT handling out into its own activity").

Intents that come from shortcuts will be seen as coming from the
launcher, so using a permission to protect shortcut intents is a
non-starter (we cannot expect launchers to add arbitrary permissions to
their manifests).  Reverting the change to intent handling isn't a
particularly palatable option, since it leaks privileges and exposes
user data to any other apps running on the device.

Instead, we add a new RUN_SHORTCUT intent (and RunShortcut activity to
handle it) which requires shortcuts to be encrypted and authenticated
using randomly-generated keys known only to us.  This ensures that only
shortcuts that the user has created can be run, and that the contents of
those shortcuts (i.e. the commands that are being run) cannot be
discovered by someone examining stored shortcuts or intent traffic on
the system.  (It is possible for an attacker who captures a sent intent
to replay that intent, repeating the command; the utility of this to an
attacker is dubious, since he cannot control, or even discover, what
command he's triggering.)

Existing user shortcuts must be recreated.  This is unavoidable, given
that existing shortcuts are hard-coded to use the RUN_SCRIPT intent via
RemoteInterface, which is something that we can no longer support.